optimization

+++ RayTracerWithCuda/GlobalTypedef.h
@@ -9,8 +9,8 @@ using unt16 = short;
 using uint16 = unsigned short;
 using int32 = int;
 using uint32 = unsigned int;
-using int64 = long;
-using uint64 = unsigned long;
+using int64 = long long;
+using uint64 = unsigned long long;
 
 using String = std::string;
